+++ pandas/tests/indexes/period/test_period.py
@@ -272,15 +272,6 @@ class TestPeriodIndex:
         exp = Index([x.ordinal for x in index])
         tm.assert_index_equal(result, exp)
 
-    def test_format_empty(self):
-        # GH35712
-        empty_idx = PeriodIndex([], freq="Y")
-        msg = r"PeriodIndex\.format is deprecated"
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            assert empty_idx.format() == []
-        with tm.assert_produces_warning(FutureWarning, match=msg):
-            assert empty_idx.format(name=True) == [""]
-
     def test_period_index_frequency_ME_error_message(self):
         msg = "Invalid frequency: 2ME"
